\docType{data}
\name{dataArray}
\alias{dataArray}
\title{Belgian First Names Count of 2013}
\source{
http://statbel.fgov.be/fr/modules/publications/statistiques/population/prenoms_de_la_population_totale.jsp
}
\description{
It covers the first names of 10.573.205 people living in
Belgium in 2013 (the Belgian population was of 11.116.243
in 2012).
}
\examples{
\dontrun{
data(prenomsBe2013)
str(dataArray)
}
}

