\name{firstNamesBe-package}
\alias{firstNamesBe}
\docType{package}
\title{
Belgian statistics based on first names.
}
\description{
This package contains function to get stats based on first names as an input. It is based on a dataset of (almost) all Belgian first names in 2013.
For any given first name, the user gets the split by gender, region and age range. If the user knows any of those 3 dimensions, he can also input them and will get an output restricted to these values.
}
\details{
\tabular{ll}{
Package: \tab firstNamesBe\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-05\cr
License: \tab MIT\cr
}

}
\author{
Nassim Haddad

Maintainer: Nassim Haddad <nassim.hd@gmail.com>
}
\examples{
help(nameStats)
}
